+++ python/api/main/resource/account.py
@@ -1,5 +1,7 @@
 import json
 
+from decimal import Decimal
+
 from flask import abort,make_response
 from flask_restful import fields,marshal,reqparse,Resource
 
@@ -76,7 +78,7 @@ class AccountApi(Resource):
         # set the arguments for the request
         parser = reqparse.RequestParser()
         parser.add_argument('name', required=True)
-        parser.add_argument('balance', type=float, default=0)
+        parser.add_argument('balance', type=Decimal, default=0)
         parser.add_argument('type', choices=('DEBIT', 'CREDIT'), default='DEBIT')
         args = parser.parse_args()
 
@@ -100,7 +102,7 @@ class AccountApi(Resource):
         # set the arguments for the request
         parser = reqparse.RequestParser()
         parser.add_argument('name')
-        parser.add_argument('balance', type=float)
+        parser.add_argument('balance', type=Decimal)
         parser.add_argument('type', choices=('DEBIT', 'CREDIT'))
         args = parser.parse_args()
